change to float type for the division

+++ schema/planta_power.sql
@@ -3,6 +3,6 @@ create table planta_power
     id              bigint auto_increment,
     id_i            int not null,
     active_power_im int not null,
-    division        int not null,
+    division        float not null,
     primary key (id)
 );
